requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.2.0
alembic==1.13.3
asttokens==2.4.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
debugpy==1.6.7
decorator==5.1.1
diffusers==0.30.3
dill==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
fonttools==4.53.1
freetype-py==2.3.0
frozenlist==1.4.0
fsspec==2024.6.1
fuzzywuzzy==0.18.0
greenlet==3.1.1
h2==4.1.0
hpack==4.0.0
huggingface_hub==0.24.6
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Levenshtein==0.26.0
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
nest_asyncio==1.6.0
networkx==3.2.1
numpy==2.1.1
optuna==4.0.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycairo==1.27.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
pytdc
python-dateutil==2.9.0
python-Levenshtein==0.26.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.10.0
rdkit
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
rlPyCairo==0.2.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
stack-data==0.6.2
statsmodels==0.14.4
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
xxhash==2.0.2
yarl==1.11.0
zipp==3.20.2
zstandard==0.23.0
